+++ NEWS.md
@@ -7,6 +7,7 @@
 * Rename variables named type that is not acutually of type DataTypes to avoid confusion
 * Scattered the functions from the utils.jl and GUIutils.jl files onto smaller files in the folders `src/utils_gen` and `src/utils_GUI`
 * Updated docstrings
+* Change variable name from `node` to `element` for variables of type `Plotable`
 
 ## Version 0.5.0 (2024-04-19)